Update vyos-modular to support building saggita (current)
